+++ GRDB Tests/Public/Core/DatabaseValueConvertible/RawRepresentableTests.swift
@@ -1,6 +1,18 @@
 import XCTest
 import GRDB
 
+enum Color32 : Int32 {
+    case Red
+    case White
+    case Rose
+}
+
+enum Color64 : Int64 {
+    case Red
+    case White
+    case Rose
+}
+
 enum Color : Int {
     case Red
     case White
@@ -13,6 +25,8 @@ enum Grape : String {
     case Riesling
 }
 
+extension Color32 : DatabaseInt32Representable { }
+extension Color64 : DatabaseInt64Representable { }
 extension Color : DatabaseIntRepresentable { }
 extension Grape : DatabaseStringRepresentable { }
 
@@ -27,6 +41,78 @@ class RawRepresentableTests: GRDBTestCase {
         }
         assertNoError {
             try migrator.migrate(dbQueue)
+        }
+    }
+    
+    func testColor32() {
+        assertNoError {
+            try dbQueue.inTransaction { db in
+                
+                do {
+                    for color in [Color32.Red, Color32.White, Color32.Rose] {
+                        try db.execute("INSERT INTO wines (color) VALUES (?)", arguments: [color])
+                    }
+                    try db.execute("INSERT INTO wines (color) VALUES (?)", arguments: [4])
+                    try db.execute("INSERT INTO wines (color) VALUES (NULL)")
+                }
+                
+                do {
+                    let rows = Row.fetchAll(db, "SELECT color FROM wines ORDER BY color")
+                    let colors = rows.map { $0.value(atIndex: 0) as Color32? }
+                    XCTAssertTrue(colors[0] == nil)
+                    XCTAssertEqual(colors[1]!, Color32.Red)
+                    XCTAssertEqual(colors[2]!, Color32.White)
+                    XCTAssertEqual(colors[3]!, Color32.Rose)
+                    XCTAssertTrue(colors[4] == nil)
+                }
+                
+                do {
+                    let colors = Optional<Color32>.fetchAll(db, "SELECT color FROM wines ORDER BY color")
+                    XCTAssertTrue(colors[0] == nil)
+                    XCTAssertEqual(colors[1]!, Color32.Red)
+                    XCTAssertEqual(colors[2]!, Color32.White)
+                    XCTAssertEqual(colors[3]!, Color32.Rose)
+                    XCTAssertTrue(colors[4] == nil)
+                }
+                
+                return .Rollback
+            }
+        }
+    }
+    
+    func testColor64() {
+        assertNoError {
+            try dbQueue.inTransaction { db in
+                
+                do {
+                    for color in [Color64.Red, Color64.White, Color64.Rose] {
+                        try db.execute("INSERT INTO wines (color) VALUES (?)", arguments: [color])
+                    }
+                    try db.execute("INSERT INTO wines (color) VALUES (?)", arguments: [4])
+                    try db.execute("INSERT INTO wines (color) VALUES (NULL)")
+                }
+                
+                do {
+                    let rows = Row.fetchAll(db, "SELECT color FROM wines ORDER BY color")
+                    let colors = rows.map { $0.value(atIndex: 0) as Color64? }
+                    XCTAssertTrue(colors[0] == nil)
+                    XCTAssertEqual(colors[1]!, Color64.Red)
+                    XCTAssertEqual(colors[2]!, Color64.White)
+                    XCTAssertEqual(colors[3]!, Color64.Rose)
+                    XCTAssertTrue(colors[4] == nil)
+                }
+                
+                do {
+                    let colors = Optional<Color64>.fetchAll(db, "SELECT color FROM wines ORDER BY color")
+                    XCTAssertTrue(colors[0] == nil)
+                    XCTAssertEqual(colors[1]!, Color64.Red)
+                    XCTAssertEqual(colors[2]!, Color64.White)
+                    XCTAssertEqual(colors[3]!, Color64.Rose)
+                    XCTAssertTrue(colors[4] == nil)
+                }
+                
+                return .Rollback
+            }
         }
     }